split up list-based GroupSerializers into its own files. update tests. update references
