+++ src/sentry/api/endpoints/organization_group_index.py
@@ -20,8 +20,7 @@ from sentry.api.helpers.group_index import (
     update_groups,
 )
 from sentry.api.paginator import DateTimePaginator, Paginator
-from sentry.api.serializers import serialize
-from sentry.api.serializers.models.group import StreamGroupSerializerSnuba
+from sentry.api.serializers import StreamGroupSerializerSnuba, serialize
 from sentry.api.utils import InvalidParams, get_date_range_from_stats_period
 from sentry.constants import ALLOWED_FUTURE_DELTA
 from sentry.exceptions import InvalidSearchQuery
