+++ src/sentry/api/endpoints/organization_group_index_stats.py
@@ -5,8 +5,7 @@ from rest_framework.response import Response
 from sentry.api.bases import OrganizationEventPermission, OrganizationEventsEndpointBase
 from sentry.api.endpoints.organization_group_index import ERR_INVALID_STATS_PERIOD
 from sentry.api.helpers.group_index import build_query_params_from_request, calculate_stats_period
-from sentry.api.serializers import serialize
-from sentry.api.serializers.models.group import StreamGroupSerializerSnuba
+from sentry.api.serializers import StreamGroupSerializerSnuba, serialize
 from sentry.api.utils import InvalidParams, get_date_range_from_stats_period
 from sentry.models import Group
 from sentry.types.ratelimit import RateLimit, RateLimitCategory
