+++ src/sentry/api/endpoints/organization_user_issues.py
@@ -4,8 +4,7 @@ from rest_framework.response import Response
 from sentry import tagstore
 from sentry.api.base import EnvironmentMixin
 from sentry.api.bases.organization import OrganizationEndpoint
-from sentry.api.serializers import serialize
-from sentry.api.serializers.models.group import TagBasedStreamGroupSerializer
+from sentry.api.serializers import TagBasedStreamGroupSerializer, serialize
 from sentry.models import EventUser, Group, ProjectTeam, Team
